Added a lisence to the project.
I chose the CC0 lisence because Rblx has some really and it would remove the legal trouble and some weird edge cases. Like for example if someone wants to copy the code to Rblx from the repo they would techically ask you and othercontributors for permission to upload as the project is currently all rights reserved. Also all free models on Rblx are defacto public domain and the CC0 lisence is public domain.